Refactor job classes into new jobs package
