+++ glacium/recipes/prep.py
@@ -1,9 +1,9 @@
 """Recipe preparing meshes and conversions before solver runs."""
 
 from glacium.managers.RecipeManager import RecipeManager, BaseRecipe
-from glacium.engines.xfoil_jobs import XfoilRefineJob, XfoilThickenTEJob
+from glacium.jobs.xfoil_jobs import XfoilRefineJob, XfoilThickenTEJob
 from glacium.engines.xfoil_convert_job import XfoilConvertJob
-from glacium.engines.pointwise_jobs import PointwiseGCIJob
+from glacium.jobs.pointwise_jobs import PointwiseGCIJob
 from glacium.engines.fluent2fensap import Fluent2FensapJob
