+++ glacium/recipes/solver.py
@@ -1,7 +1,7 @@
 """Recipe executing the FENSAP solver chain."""
 
 from glacium.managers.RecipeManager import RecipeManager, BaseRecipe
-from glacium.engines.fensap import FensapRunJob, Drop3dRunJob, Ice3dRunJob
+from glacium.jobs.fensap_jobs import FensapRunJob, Drop3dRunJob, Ice3dRunJob
 
 
 @RecipeManager.register
